'warmup added'

+++ C/src/CHeterodyning.c
@@ -1,7 +1,7 @@
 #include "CHeterodyning.h"
 
-extern float data[SAMPLE_COUNT];
-extern float carrier[SAMPLE_COUNT];
+extern __fp16 data[SAMPLE_COUNT];
+extern __fp16 carrier[SAMPLE_COUNT];
 
 float result [SAMPLE_COUNT];
